+++ skuhub/src/main/java/org/skuhub/skuhub/repository/timetable/CompletionRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface CompletionRepository extends JpaRepository<CompletionEntity, Integer> {
-    Optional<CompletionEntity> findByUserInfoTb_UserKey(Integer userKey);
+    Optional<CompletionEntity> findByUserInfoTb_UserKey(Long userKey);
 }
